Improve exception handling in forward_function.
If we throw string, then StackTrace generation is not performed and v8::TryCatch::StackTrace() returns Empty maybe. While v8::TryCatch::Exception() returns exception string as expected.
So it's better to throw v8::Error, v8::TypeError etc, so we can easily track root of the problem.

By the way it seems that calling args.GetReturnValue().Set(ex) is not nessesary to handle exceptions. So it seems it can be removed.